refactor: 클래스 디렉터리 구조 변경
