+++ src/main/java/lotto/domain/model/LottoNumber.java
@@ -1,4 +1,4 @@
-package domain.lotto;
+package lotto.domain.model;
 
 public class LottoNumber {
