+++ src/main/java/lotto/domain/model/Lottos.java
@@ -1,4 +1,4 @@
-package domain.lotto;
+package lotto.domain.model;
 
 import java.util.Collections;
 import java.util.List;
